+++ PlantUmlViewer/PlantUmlViewer/DiagramGeneration/DiagramGenerator.cs
@@ -26,7 +26,7 @@ namespace PlantUmlViewer.DiagramGeneration
             this.plantUmlBinary = plantUmlBinary;
         }
 
-        public async Task<List<GeneratedDiagram>> GenerateDocumentAsync(string text, CancellationTokenSource cancellationTokenSource)
+        public async Task<List<GeneratedDiagram>> GenerateDocumentAsync(string text, string include, CancellationTokenSource cancellationTokenSource)
         {
             /*
              * The PlantUML responses could contain multiple pages, where each page could contain the images for multiple diagrams
@@ -51,13 +51,13 @@ namespace PlantUmlViewer.DiagramGeneration
             //Generate the first page directly at startup
             List<Task<bool>> generateTasks = new List<Task<bool>>()
             {
-                Task.Run(() => GeneratePageAsync(text, 0, pages, cancellationTokenSource))
+                Task.Run(() => GeneratePageAsync(text, include, 0, pages, cancellationTokenSource))
             };
             //Generate the (maybe) following pages
             int pageIndex = 1;
             while (true)
             {
-                generateTasks.Add(Task.Run(() => GeneratePageAsync(text, pageIndex, pages, cancellationTokenSource)));
+                generateTasks.Add(Task.Run(() => GeneratePageAsync(text, include, pageIndex, pages, cancellationTokenSource)));
                 await Task.WhenAll(generateTasks);
                 if (generateTasks.Any(rT => !rT.Result))
                 {
@@ -74,7 +74,8 @@ namespace PlantUmlViewer.DiagramGeneration
             return ReorganizePagesToDiagram(pages);
         }
 
-        private async Task<bool> GeneratePageAsync(string text, int pageIndexToGenerate, Dictionary<int, Dictionary<int, SvgDocument>> pages, CancellationTokenSource cancellationTokenSource)
+        private async Task<bool> GeneratePageAsync(string text, string include, int pageIndexToGenerate,
+            Dictionary<int, Dictionary<int, SvgDocument>> pages, CancellationTokenSource cancellationTokenSource)
         {
             IPlantUmlRenderer renderer = renderFactory.CreateRenderer(new PlantUmlSettings()
             {
@@ -82,6 +83,7 @@ namespace PlantUmlViewer.DiagramGeneration
                 LocalPlantUmlPath = plantUmlBinary,
                 JavaPath = javaPath,
                 RenderingMode = RenderingMode.Local,
+                Include = include,
                 Delimitor = DIAGRAM_DELIMITOR,
                 ImageIndex = pageIndexToGenerate
             });