+++ PlantUmlViewer/PlantUmlViewer/Settings/PlantUmlViewerSettings.cs
@@ -3,6 +3,7 @@
     public class PlantUmlViewerSettings
     {
         public string JavaPath { get; set; } = "";
+        public string Include { get; set; } = "";
         public decimal ExportSizeFactor { get; set; } = 1;
     }
 }